trakt/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
peewee==3.14.8
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
